feat(TheaterAcceptanceTest): Divide admin and basic
